Draw the intensity of Hawkes Process

Draw the intensity of Hawkes Process, a helper function for 'drawUniMMHPIntensity' while also available independently

drawHPIntensity(
  hp,
  start = 0,
  end = max(events),
  history = 0,
  events,
  color = 1,
  i = 1,
  add = FALSE,
  fit = FALSE,
  plot_events = FALSE,
  vec = NULL,
  int_title = "Hawkes Intensity"
)

Arguments

hp
object parameters for Hawkes process

start
the start time of current state

end
the end time of current state

history
the past event times

events
the event times happened in this state

color
A specification for the default plotting color.

i
state number this corresponds to a state jump directly before which is only important when using mmhp

add
whether to add the hawkes intensity to an existing plot

fit
a boolean indicating whether to fit a hp or use the passed object

plot_events
a boolean indicating whether events inputted will be plotted

vec
vector of initial values of parameters used in fithp

int_title
title of the intensity plot

Examples

if (FALSE) {
hp_obj <- hp(lambda0 = 0.1, alpha = 0.45, beta = 0.5)
sims <- simulatehp(hp_obj, start = 0, end = 20, history = 0)
events <- sims$events
drawHPIntensity(hp_obj, start = 0, end = max(events), history = 0, events)
}